move maxbound update logic to base component
